+++ HISTORY.md
@@ -2,7 +2,7 @@
 
 ## in development:
 * **New Features:**
-    * include data dependencies in libraries, eliminates the need for initialization routines creating and populating variables required by custom blocks
+    * include custom block data dependencies in libraries, eliminates the need for initialization routines creating and populating variables required by custom block definitions
     * visually highlight variable scope when mouse-hovering over a variable declaration in visible stepping mode
     * visually highlight the variable declaration of variable accessor blocks when mouse-hovering over them in visible stepping mode
     * visually highlight senders / receivers when mouse-hovering over messaging blocks in visible stepping mode
@@ -89,8 +89,9 @@
 * gui, store: read back data dependencies included in exported libraries
 * byob, gui: distinguish between global and sprite-local data dependencies
 * gui, store: read back local data dependencies included in exported libraries
-* blocks, gui: include data dependencies in exported single scripts
+* blocks, gui: include custom block data dependencies in exported single scripts
 * blocks: only include data dependencies in exported block definitions
+* objects, gui: include global custom block data dependencies in exported sprites
 
 ### 2023-10-22
 * byob, gui, threads: new file structure for libraries to include global data dependencies